refactor: Moved transient-skill to new directory
