+++ presentation/sheet/item/skill/skill-chat-message-viewmodel.mjs
@@ -1,4 +1,4 @@
-import TransientSkill, { SKILL_HEAD_STATES } from "../../../../business/document/item/transient-skill.mjs"
+import TransientSkill, { SKILL_HEAD_STATES } from "../../../../business/document/item/skill/transient-skill.mjs"
 import { ATTRIBUTES } from "../../../../business/ruleset/attribute/attributes.mjs"
 import { validateOrThrow } from "../../../../business/util/validation-utility.mjs"
 import ButtonToggleVisibilityViewModel from "../../../component/button-toggle-visibility/button-toggle-visibility-viewmodel.mjs"
